requirements.txt
mysql-connector-python
customtkinter
flake8
pytest
python-dotenv
logging
colorama
pytz